requirements.txt
google-generativeai
python-dotenv
ollama
PyYAML
jellyfish
python-Levenshtein
scipy
tqdm
scikit-learn
pandas
langchain
openai
tiktoken